requirements-webhook.txt
# Minimal dependencies for webhook server only
# This is for Railway deployment to keep image size under 4GB

# Core web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
supabase>=2.18.0,<3.0.0
asyncpg>=0.29.0

# Airtable
pyairtable>=2.3.0

# Article scraping (lightweight)
newspaper3k>=0.2.8
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Configuration
pydantic>=2.11.7,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
aiohttp>=3.9.0
